# Improve perf and scalability of Regex's cache