… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Reference.cs
… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/Regex.Cache.cs