Regex maintains a cache used for the static methods on Regex, e.g. Regex.IsMatch.  The cache is implemented as an LRU cache, which maintains a linked list and a dictionary of the cached instances.  The linked list maintains the order in which the cached instances were last accessed, making it cheap to expunge older items from the cache.  However, that comes at a significant cost: unless the item is the very first one in the linked list, all reads on the cache require taking a global lock, because the linked list needs to be mutated to move the found node to the beginning.  That lock has both throughput and scalability implications.

This PR changes the cache from using a `Dictionary<>` and a linked list to instead using a `ConcurrentDictionary<>` and a `List<>`.  Rather than making all accesses more expensive in order to make drops less expensive, it makes all reads much cheaper and more scalable, at the expense of making drops more expensive.  Since dropping from the cache means we're already paying the expensive cost of creating/parsing/compiling/etc. a new Regex instance, this is a better trade-off, especially since any frequent dropping suggests the consuming app or library needs to revisit its Regex strategy, either using Regex.CacheSize to increase the cache size appropriately, or doing its own caching (e.g. creating the Regex instance it needs and storing it into a field for all future use).

The new scheme uses a `ConcurrentDictionary<Key,Node>`, a `List<Node>`, and a fast-path field storing the most recently used Regex instance (just as the existing implementation did).  On lookups, if the fast-path field has the matching value, it's just returned.  Otherwise, the dictionary is consulted, and if the item is found, the fast-path field is updated.  No locking at all is employed, and only a few volatile read/writes are used to update a "last access stamp" that's used to indicate importance if/when items do need to be expunged.  On additions, we do still take a global lock and add to the cache.  If this puts us over our cache size, we pick an item from the list and remove it.  If the list is small, we just examine all of the items looking for the oldest.  If the list is larger, we examine a random subset of it; we may not get rid of the absolute oldest item, but it'll be old enough.

cc: @danmosemsft, @eerhardt, @ViktorHofer 

Results from running master (old) vs this PR (new) on the *Regex*Cache* tests from the dotnet/performance repo:

|                 Method |        Toolchain |  total | unique | cacheSize |       Mean | Ratio |      Gen 0 |
|----------------------- |----------------- |------- |------- |---------- |-----------:|------:|-----------:|
|                IsMatch | new              |  40000 |      7 |         0 |  55.792 ms |  0.96 | 24750.0000 |
|                IsMatch | old              |  40000 |      7 |         0 |  57.454 ms |  1.00 | 25000.0000 |
|                        |                  |        |        |           |            |       |            |
| IsMatch_Multithreading | new              |  40000 |      7 |         0 |  31.867 ms |  0.98 | 25000.0000 |
| IsMatch_Multithreading | old              |  40000 |      7 |         0 |  33.059 ms |  1.00 | 25000.0000 |
|                        |                  |        |        |           |            |       |            |
|                IsMatch | new              |  40000 |   1600 |        15 | 112.867 ms |  0.64 | 38000.0000 |
|                IsMatch | old              |  40000 |   1600 |        15 | 176.183 ms |  1.00 | 39000.0000 |
|                        |                  |        |        |           |            |       |            |
| IsMatch_Multithreading | new              |  40000 |   1600 |        15 |  54.960 ms |  0.50 | 38000.0000 |
| IsMatch_Multithreading | old              |  40000 |   1600 |        15 | 109.076 ms |  1.00 | 39000.0000 |
|                        |                  |        |        |           |            |       |            |
|                IsMatch | new              |  40000 |   1600 |       800 |  87.606 ms |  0.51 | 20000.0000 |
|                IsMatch | old              |  40000 |   1600 |       800 | 174.088 ms |  1.00 | 22000.0000 |
|                        |                  |        |        |           |            |       |            |
| IsMatch_Multithreading | new              |  40000 |   1600 |       800 |  50.726 ms |  0.41 | 20000.0000 |
| IsMatch_Multithreading | old              |  40000 |   1600 |       800 | 123.640 ms |  1.00 | 22000.0000 |
|                        |                  |        |        |           |            |       |            |
|                IsMatch | new              |  40000 |   1600 |      3200 |  13.444 ms |  0.94 |          - |
|                IsMatch | old              |  40000 |   1600 |      3200 |  14.247 ms |  1.00 |          - |
|                        |                  |        |        |           |            |       |            |
| IsMatch_Multithreading | new              |  40000 |   1600 |      3200 |   5.500 ms |  0.42 |          - |
| IsMatch_Multithreading | old              |  40000 |   1600 |      3200 |  13.180 ms |  1.00 |          - |
|                        |                  |        |        |           |            |       |            |
|                IsMatch | new              | 400000 |      1 |        15 |  41.607 ms |  1.00 |          - |
|                IsMatch | old              | 400000 |      1 |        15 |  41.512 ms |  1.00 |          - |
|                        |                  |        |        |           |            |       |            |
| IsMatch_Multithreading | new              | 400000 |      1 |        15 |  40.066 ms |  0.90 | 18000.0000 |
| IsMatch_Multithreading | old              | 400000 |      1 |        15 |  44.558 ms |  1.00 | 33500.0000 |
|                        |                  |        |        |           |            |       |            |
|                IsMatch | new              | 400000 |      7 |        15 |  66.953 ms |  0.93 |          - |
|                IsMatch | old              | 400000 |      7 |        15 |  71.789 ms |  1.00 |          - |
|                        |                  |        |        |           |            |       |            |
| IsMatch_Multithreading | new              | 400000 |      7 |        15 |  46.878 ms |  0.52 | 12000.0000 |
| IsMatch_Multithreading | old              | 400000 |      7 |        15 |  90.335 ms |  1.00 |  9000.0000 |