@stephentoub 2019-12-05
The CI failures here are strange; lots of EventSource tests failing with, e.g.
```
    BasicEventSourceTests.TestsWrite.Test_Write_T_ETW [FAIL]
      Assert.Equal() Failure
                 (pos 0)
      Expected: 
      Actual:   System.Collections.Concurrent.ConcurrentCúúú
                 (pos 0)
      Stack Trace:
        /_/src/libraries/System.Diagnostics.Tracing/tests/BasicEventSourceTest/TestUtilities.cs(50,0): at BasicEventSourceTests.TestUtilities.CheckNoEventSourcesRunning(String message)
        /_/src/libraries/System.Diagnostics.Tracing/tests/BasicEventSourceTest/TestsWrite.cs(456,0): at BasicEventSourceTests.TestsWrite.Test_Write_T(Listener listener)
        /_/src/libraries/System.Diagnostics.Tracing/tests/BasicEventSourceTest/TestsWrite.Etw.cs(29,0): at BasicEventSourceTests.TestsWrite.Test_Write_T_ETW()
```
My assumption is that a) there was some kind of change in coreclr recently that is now causing a discrepancy with the tests, and b) this is now showing up after @safern's live/live change went in last night, but I'm not sure why I don't see similar failures on other PRs, nor why the "Actual" string above looks corrupted ("System.Collections.Concurrent.ConcurrentCúúú").  Regardless, I put up https://github.com/dotnet/runtime/pull/565 to add this EventSource to the test's exempted list. @noahfalk, ideas?