@safern 2019-12-05
> but I'm not sure why I don't see similar failures on other PRs.

Does this repro locally with and without your change?