@stephentoub 2019-12-06
> Does this repro locally with and without your change?

No.