@danmosemsft 2020-01-06
Maybe `úúú` is meant to be an ellipsis with some special period, and the console codepage is corrupting it.